+++ packages/notice-bar/index.ts
@@ -1,6 +1,6 @@
 import { VantComponent } from '../common/component';
 import { Weapp } from 'definitions/weapp';
-import { requestAnimationFrame } from '../common/utils';
+import { getRect, requestAnimationFrame } from '../common/utils';
 
 VantComponent({
   props: {
@@ -70,9 +70,9 @@ VantComponent({
   methods: {
     init() {
       Promise.all([
-        this.getRect('.van-notice-bar__content'),
-        this.getRect('.van-notice-bar__wrap'),
-      ]).then((rects: WechatMiniprogram.BoundingClientRectCallbackResult[]) => {
+        getRect.call(this, '.van-notice-bar__content'),
+        getRect.call(this, '.van-notice-bar__wrap'),
+      ]).then((rects) => {
         const [contentRect, wrapRect] = rects;
         if (
           contentRect == null ||
@@ -85,7 +85,7 @@ VantComponent({
 
         const { speed, scrollable, delay } = this.data;
 
-        if (scrollable && wrapRect.width < contentRect.width) {
+        if (scrollable || wrapRect.width < contentRect.width) {
           const duration = (contentRect.width / speed) * 1000;
 
           this.wrapWidth = wrapRect.width;